fix: Move verifyOsProberBlockEmpty to TestCustomizeImageNewUUIDs
